% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Performs the bootstrap procedure to make a confidence interval for the specified statistics of a variable
Plots a histogram of the bootstrap values with the specified confidence interval cutoffs}
\usage{
bootstrap(data, nreps, var, stats = mean, lower = 0.025, upper = 0.975)
}
\arguments{
\item{data}{a data frame}

\item{nreps}{the number of bootstraps that are to be generated}

\item{var}{the variable name for the column in the data set}

\item{stats}{the statistic used to perform the bootstrap procedure}

\item{lower}{the lower quantile of the confidence interval}

\item{upper}{the upper quantile of the confidence interval}
}
\value{
the confidence interval and a histogram of the bootstrap values with the confidence interval cutoffs
}
\description{
Performs the bootstrap procedure to make a confidence interval for the specified statistics of a variable
Plots a histogram of the bootstrap values with the specified confidence interval cutoffs
}
