% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfolds_cross_valid.R
\name{kfolds_cross_valid}
\alias{kfolds_cross_valid}
\title{Performs the K-folds cross validation procedure and returns the mean square errors}
\usage{
kfolds_cross_valid(K, data, y, x)
}
\arguments{
\item{K}{the number of folds}

\item{data}{a data frame}

\item{y}{the response variable for the linear regression}

\item{x}{a vector of the explanatory variable(s) for the linear regression}
}
\value{
a vector of average mean square errors
}
\description{
Performs the K-folds cross validation procedure and returns the mean square errors
}
