% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{permutation_test}
\alias{permutation_test}
\title{Performs a permutation test and returns a two-tail p-value
Plots a histogram of the differences in the specified sample statistic}
\usage{
permutation_test(data, cat_var, quant_var, nreps, stats = mean)
}
\arguments{
\item{data}{a data frame}

\item{cat_var}{the categorical variable of interest}

\item{quant_var}{the quantitative variable of interest}

\item{nreps}{the number of permutations}

\item{stats}{the statistic used to calculate the data}
}
\value{
a two-tail p-value and a histrogram of the differences in the specified sample statistic
}
\description{
Performs a permutation test and returns a two-tail p-value
Plots a histogram of the differences in the specified sample statistic
}
